Add Rock Paper Scissors game
Game is built on an abstraction for verifying repeated simultaneous games.
This is a powerful class of games because it can model any case where only
the state, not the program, changes between rounds of the game. Simultaneous
games can emulate sequential ones by alternating players, so many games are
playable using this framework.

Two library contracts, RpsState and RpsAction, are used to abstract the
state and action. This allows the type to be changed from bytes32 to a
larger fixed array (or content-addressed storage hash) by modifying the
libraries without refactoring the code.